requirements.txt
asgiref==3.8.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
fonttools==4.57.0
gunicorn==23.0.0
kiwisolver==1.4.8
lab==8.4
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
report==0.0.1
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
txt2tags==3.9
tzdata==2025.2
whitenoise==6.9.0
reportlab==4.4.0
PyPDF2==3.0.1

